processor/requirements.txt
Pillow==10.0.0
rembg==2.0.50
watchdog==3.0.0
python-dotenv==1.0.0
pyyaml==6.0.1
redis==5.0.1
